--- src/features/create-dao/steps/DAOSummary.module.scss
@@ -1,3 +1,0 @@
-.Footer {
-	margin-top: 1rem;
-}
