--- src/features/create-dao/steps/GovernanceForm.module.scss
@@ -1,3 +1,0 @@
-.Footer {
-	margin-top: 5rem;
-}
