--- src/features/create-dao/steps/index.ts
@@ -1,4 +1,0 @@
-export * from './DAOSummary';
-export * from './DetailsForm';
-export * from './GovernanceForm';
-export * from './TreasuryForm';
